renamed get_M_r_w -> get_sign_times_M_r_w

+++ include/dca/phys/dca_step/cluster_solver/shared_tools/accumulation/mc_accumulator_data.hpp
@@ -27,7 +27,7 @@ public:
     return accumulated_sign;
   }
 
-  double get_number_of_measurements() const {
+  int get_number_of_measurements() const {
     return number_of_measurements;
   }
